Commit Jesus 28-09

+++ HiberLibros/src/main/java/com/hiberlibros/HiberLibros/services/AutorService.java
@@ -10,9 +10,12 @@ import org.springframework.stereotype.Service;
 
 import com.hiberlibros.HiberLibros.dtos.LibroDto;
 import com.hiberlibros.HiberLibros.entities.Autor;
+import com.hiberlibros.HiberLibros.entities.Libro;
 import com.hiberlibros.HiberLibros.interfaces.IAutorService;
+import com.hiberlibros.HiberLibros.interfaces.ILibroService;
 import com.hiberlibros.HiberLibros.repositories.AutorLibroRepository;
 import com.hiberlibros.HiberLibros.repositories.AutorRepository;
+import java.util.ArrayList;
 
 @Service
 public class AutorService implements IAutorService {
@@ -22,18 +25,22 @@ public class AutorService implements IAutorService {
 
     @Autowired
     private AutorLibroRepository repo;
+    
+    @Autowired
+    private ILibroService serviceLib;
 
     @Autowired
     private ModelMapper obj;
 
     @Override
     public List<Autor> buscarAutores(String buscar) {
-        return autorRepo.findAll().stream()
-                .filter(
-                        x -> x.getNombre().concat(x.getApellidos()).toLowerCase()//paso nombre completo a minusuclas para comparar
-                                .contains(buscar.toLowerCase())
-                )
-                .collect(Collectors.toList());
+        return autorRepo.findByNombreContainsIgnoreCaseAndApellidosContainsIgnoreCase(buscar, buscar);
+//        return autorRepo.findAll().stream()
+//                .filter(
+//                        x -> x.getNombre().concat(x.getApellidos()).toLowerCase()//paso nombre completo a minusuclas para comparar
+//                                .contains(buscar.toLowerCase())
+//                )
+//                .collect(Collectors.toList());
     }
 
     @Override
@@ -52,17 +59,32 @@ public class AutorService implements IAutorService {
 
     @Override
     public Optional<Autor> encontrarAutor(Integer id) {
-        return autorRepo.findById(id);
+        return autorRepo.findByIdAutorAndDesactivado(id, Boolean.FALSE);
     }
 
     @Override
     public void borrarAutor(Integer id) {
-        autorRepo.deleteById(id);
+        Optional<Autor> a=encontrarAutor(id);
+        List<Libro> l=new ArrayList<>();
+        if(a.isPresent()){
+            l=serviceLib.encontrarPorAutor(a.get());
+             if(l.size()==0 || l==null){
+                 autorRepo.deleteById(id);
+             }else{
+                 a.get().setDesactivado(Boolean.TRUE);
+                 autorRepo.save(a.get());
+                 l.forEach(x->{
+                     serviceLib.bajaLibroId(x.getId());
+                 });
+             }
+        }
+
+        
 
     }
 
     @Override
     public List<Autor> consultarAutores() {
-        return autorRepo.findAll();
+        return autorRepo.findByDesactivado(Boolean.FALSE);
     }
 }
